+++ lib/logger/error_handler.ex
@@ -1,10 +1,11 @@
 defmodule Logger.ErrorHandler do
   use GenEvent
 
-  # TODO: Make threshold configurable
-  def init({otp?, sasl?}) do
-    {:ok, %{otp: otp?, sasl: sasl?, threshold: 50,
-            messages: 0, last_time: :os.timestamp, dropped: 0}}
+  require Logger
+
+  def init({otp?, sasl?, threshold}) do
+    {:ok, %{otp: otp?, sasl: sasl?, threshold: threshold,
+            last_length: 0, last_time: :os.timestamp, dropped: 0}}
   end
 
   ## Handle event
@@ -44,32 +45,36 @@ defmodule Logger.ErrorHandler do
     state
   end
 
-  defp check_threshold(%{messages: messages, threshold: threshold} = state)
-      when messages <= threshold do
-    %{state | messages: messages + 1}
-  end
+  defp check_threshold(%{last_time: last_time, last_length: last_length,
+                         dropped: dropped, threshold: threshold} = state) do
+    {m, s, _} = current_time = :os.timestamp
+    current_length = message_queue_length()
 
-  defp check_threshold(%{last_time: last_time, dropped: dropped} = state) do
-    {m, s, _} = now = :os.timestamp
-    case last_time do
-      {^m, ^s, _} ->
-        count = drop_messages(now, 0)
-        %{state | dropped: dropped + count}
-      {_, _, _} ->
+    cond do
+      match?({^m, ^s, _}, last_time) and current_length - last_length > threshold ->
+        count = drop_messages(current_time, 0)
+        %{state | dropped: dropped + count, last_length: message_queue_length()}
+      match?({^m, ^s, _}, last_time) ->
+        state
+      true ->
         if dropped > 0 do
-          Logger.warn "Logger dropped #{dropped} OTP/SASL messages that " <>
-                      "exceeded the amount of #{state.threshold} messages/second"
+          Logger.warn "Logger dropped #{dropped} OTP/SASL messages as it " <>
+                      "exceeded the amount of #{threshold} messages/second"
         end
-        %{state | last_time: now, messages: 1, dropped: 0}
+        %{state | dropped: 0, last_time: current_time, last_length: current_length}
     end
   end
 
-  defp drop_messages(now, count) do
-    {m, s, _} = :os.timestamp
-    case now do
+  defp message_queue_length() do
+    {:message_queue_len, len} = Process.info(self(), :message_queue_len)
+    len
+  end
+
+  defp drop_messages({m, s, _} = last_time, count) do
+    case :os.timestamp do
       {^m, ^s, _} ->
         receive do
-          {:notify, _event} -> drop_messages(now, count + 1)
+          {:notify, _event} -> drop_messages(last_time, count + 1)
         after
           0 -> count
         end